healthcheck: now requires file `mounted` to be present in storage

+++ app/Http/Controllers/Api/HealthController.php
@@ -8,6 +8,10 @@ class HealthController extends ApiController
 {
     public function health()
     {
+        $montedCheckFile = storage_path('mounted');
+        if(!@file_exists($mountedCheckFile))
+            return $this->respond(503, 'storage unavailable');
+
         $healthCheckFile = storage_path('health.txt');
 
         if (@file_put_contents($healthCheckFile, date('Y-m-d H:i:s')) === false) {
